requirements.txt
numpy==1.18.1
matplotlib==3.1.3
python-dateutil==2.8.1
scikit-learn==0.22.1
scipy==1.4.1
SQLAlchemy==1.3.13
opcua==0.98.9
PyYAML==5.3
pymysql
pandas